requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg[binary]==3.2.10
dj-database-url==2.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
PyJWT==2.9.0
djangorestframework-simplejwt==5.3.1
supabase==2.4.0
drf-spectacular==0.27.2
django-filter==24.3
black==24.8.0
isort==5.13.2
flake8==7.0.0
mypy==1.11.2
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
gunicorn==22.0.0
sentry-sdk==2.14.0
django-ratelimit==4.1.0
celery==5.4.0
redis==5.0.8
Pillow==10.4.0
